Address perpetual futures review findings
