+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - **Futures review round 12**: the `PositionSizer` continuous affordability bisection gains a convergence guard that fails with `IllegalStateException` when a `Num` implementation with unbounded precision, e.g. an unbounded `MathContext` `DecimalNumFactory`, keeps the midpoint strictly between the bounds; finite-precision implementations converge far below the guard and the documented precision precondition now sits on `Context.maxAffordableAmount`. `PositionSizer` factory methods (`fixed`, `balance`, `kelly`) preserve caller amounts and principals exactly instead of normalizing them through `double`, so values beyond `2^53` survive the record initial-capital comparison unchanged for decimal number factories.
 - **Futures review round 13**: `PositionSizer` factory validation no longer narrows `BigDecimal`/`BigInteger` inputs through `doubleValue()`, and the affordability budget finiteness check uses the domain `Num.isFinite`, so exact decimal values beyond the double range are preserved and accepted; `snapshotNumber` whitelists the JDK immutable number types and snapshots every other `Number` value at creation time, so caller-owned mutable inputs can no longer change after sizer construction; the affordability search convergence guard now covers about 9800 significant digits, so a 5000-digit `DecimalNumFactory` converges instead of failing with a spurious `IllegalStateException`.
 
-- **Futures review round 14**: removed the pathological unbounded-precision affordability test from the default unit lane after the bounded 5000-digit convergence test established the supported finite-precision contract; factory Kelly validation now compares decimal inputs without double narrowing; the default-branch required status rule now points at the canonical Build and Verify gate after the retired standalone formatting workflow was consolidated; imported positions retain compatible holding-cost models and trade-index cursors, futures return accounting ignores pending fills and includes recorded extension executions, and extension-bar backtests continue until closeout; partial futures projections preserve residual open exposure, mixed number factories are normalized during fill aggregation, average-cost views retain original fill prices, modeled fees honor settlement conversions, extension-bar fee totals include executed fills, and complete stop-limit closes bypass volume caps.
+- **Futures review round 14**: removed the pathological unbounded-precision affordability test from the default unit lane after the bounded 5000-digit convergence test established the supported finite-precision contract; factory Kelly validation now compares decimal inputs without double narrowing; the default-branch required status rule now points at the canonical Build and Verify gate after the retired standalone formatting workflow was consolidated; imported positions retain compatible holding-cost models and trade-index cursors, futures return accounting ignores pending fills and includes recorded extension executions, and extension-bar backtests continue until closeout; partial futures projections preserve residual open exposure, mixed number factories are normalized during fill aggregation, average-cost views retain original fill prices, modeled fees honor settlement conversions, extension-bar fee totals include executed fills, and complete stop-limit closes bypass volume caps.; return and risk views avoid discarded-history allocations and support nullable mark prices, futures numeric hashes follow numeric equality, imported event horizons are restored, and test scenarios now live with their owning APIs.
 
 ## 0.25.0 (2026-09-07)
