+++ ta4j-core/src/test/java/org/ta4j/core/backtest/BarSeriesManagerTest.java
@@ -9,7 +9,6 @@ import static org.junit.Assert.assertThrows;
 import static org.junit.Assert.assertSame;
 import static org.junit.Assert.assertTrue;
 
-import java.lang.reflect.Method;
 import java.time.Instant;
 import java.util.ArrayList;
 import java.util.List;
@@ -740,7 +739,7 @@ public class BarSeriesManagerTest {
     }
 
     @Test
-    public void advanceToBarBoundarySkipsRawInaccessibleExtensionIndexes() throws Exception {
+    public void runOnWindowedSeriesSkipsInaccessibleExtensionIndexes() {
         BarSeries windowed = new MockBarSeriesBuilder().withNumFactory(numFactory).withMaxBarCount(4).build();
         for (int i = 0; i < 10; i++) {
             windowed.barBuilder()
@@ -751,15 +750,10 @@ public class BarSeriesManagerTest {
         assertEquals(6, windowed.getRemovedBarsCount());
         assertEquals(9, windowed.getEndIndex());
 
-        BarSeriesManager manager = new BarSeriesManager(windowed);
-        BaseTradingRecord record = BaseTradingRecord.builder().build();
-        Method boundary = BarSeriesManager.class.getDeclaredMethod("advanceToBarBoundary", TradingRecord.class,
-                int.class, boolean.class);
-        boundary.setAccessible(true);
-        // Before the raw-accessibility guard, these extension indexes threw
-        // IndexOutOfBoundsException from getBar; they are no-ops now.
-        boundary.invoke(manager, record, windowed.getEndIndex() + 1, true);
-        boundary.invoke(manager, record, windowed.getEndIndex() + 1, false);
-        boundary.invoke(manager, record, windowed.getEndIndex() + 100, true);
+        Strategy noSignalStrategy = new BaseStrategy(new FixedRule(100), new FixedRule(100));
+        TradingRecord record = new BarSeriesManager(windowed).run(noSignalStrategy);
+
+        assertTrue(record.getPositions().isEmpty());
     }
+
 }
